feat: shuffle components so that md is never used inside components
