+++ libs/core/markdown/src/compilers/compile-markdown.ts
@@ -1,0 +1,14 @@
+/**
+ * A convenience type that represents MD documents.
+ */
+export type MdDocument = {
+  mdContent: string;
+};
+
+/**
+ * Note: This doesn't do anything since for now we only compile markdown on the
+ * client.
+ */
+export const compileMarkdown = (content: string): MdDocument => ({
+  mdContent: content,
+});
