+++ libs/core/markdown/src/index.ts
@@ -1,0 +1,5 @@
+export { plugins } from "./common/plugins";
+export { compileMarkdown } from "./compilers/compile-markdown";
+export { compileMdx } from "./compilers/compile-mdx";
+export type { MdDocument } from "./compilers/compile-markdown";
+export type { MdxDocument } from "./compilers/compile-mdx";
