+++ libs/next/question-bank/src/__mocks__/annex-search.ts
@@ -1,0 +1,477 @@
+import type { AnnexSearchResult } from "@chair-flight/core/search";
+import type { AppRouterOutput } from "@chair-flight/trpc/client";
+
+const items: AnnexSearchResult[] = [
+  {
+    id: "52c43c862e9d424470b332772d96cba5",
+    href: "/content/content-question-bank-atpl/annexes/52c43c862e9d424470b332772d96cba5.jpg",
+    description: "Marshaller sign, Slow down",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "2dcb9b92b8bcef885587755725fb5fbb",
+    href: "/content/content-question-bank-atpl/annexes/2dcb9b92b8bcef885587755725fb5fbb.jpg",
+    description: "Marshaller sign, Set brakes",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "79f299f0d5d3b2f5be5d71c9dd35b709",
+    href: "/content/content-question-bank-atpl/annexes/79f299f0d5d3b2f5be5d71c9dd35b709.jpg",
+    description: "Marshaller sign, Turn Left",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "620342c09ff8c3334b4c3a10c9b502cc",
+    href: "/content/content-question-bank-atpl/annexes/620342c09ff8c3334b4c3a10c9b502cc.jpg",
+    description: "Marshaller sign, Positive",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "d3b1de1427950ff734a5c19cef4e10f7",
+    href: "/content/content-question-bank-atpl/annexes/d3b1de1427950ff734a5c19cef4e10f7.jpg",
+    description: "Marshaller sign, Straight ahead",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "a20f8674deeab521550d003d8e9aad2d",
+    href: "/content/content-question-bank-atpl/annexes/a20f8674deeab521550d003d8e9aad2d.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "9d252f5081e36717e51b5938be680344",
+    href: "/content/content-question-bank-atpl/annexes/9d252f5081e36717e51b5938be680344.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "d8d6abd464e947135f574507fc18cd97",
+    href: "/content/content-question-bank-atpl/annexes/d8d6abd464e947135f574507fc18cd97.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "32868ee6d3582c52c01473c7061579ae",
+    href: "/content/content-question-bank-atpl/annexes/32868ee6d3582c52c01473c7061579ae.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "17f0375525cb9d63c59c11d338e06f85",
+    href: "/content/content-question-bank-atpl/annexes/17f0375525cb9d63c59c11d338e06f85.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "c4337d2344f17c742000240e172212a2",
+    href: "/content/content-question-bank-atpl/annexes/c4337d2344f17c742000240e172212a2.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "eb38917314ded07e0c84adfe1a2c84c1",
+    href: "/content/content-question-bank-atpl/annexes/eb38917314ded07e0c84adfe1a2c84c1.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "63c7dab406647c2c0be0c11f1abb2224",
+    href: "/content/content-question-bank-atpl/annexes/63c7dab406647c2c0be0c11f1abb2224.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QSB6GBPB9F",
+        href: "/modules/[object Object]/questions/QSB6GBPB9F",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.02",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.02",
+      },
+    ],
+  },
+  {
+    id: "32747336da740b77e279adddc1db3d58",
+    href: "/content/content-question-bank-atpl/annexes/32747336da740b77e279adddc1db3d58.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QTNBTAFP2Q",
+        href: "/modules/[object Object]/questions/QTNBTAFP2Q",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.18",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.18",
+      },
+    ],
+  },
+  {
+    id: "a3457f89b1a18edbf4c47221df71b6ef",
+    href: "/content/content-question-bank-atpl/annexes/a3457f89b1a18edbf4c47221df71b6ef.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q5WASR1IGH",
+        href: "/modules/[object Object]/questions/Q5WASR1IGH",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.05.03.01.03",
+        href: "/modules/[object Object]/learning-objectives/010.05.03.01.03",
+      },
+    ],
+  },
+  {
+    id: "dfa38ab24245861426112235bc23848b",
+    href: "/content/content-question-bank-atpl/annexes/dfa38ab24245861426112235bc23848b.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q25JRWCYBK",
+        href: "/modules/[object Object]/questions/Q25JRWCYBK",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.05.01.06",
+        href: "/modules/[object Object]/learning-objectives/010.06.05.01.06",
+      },
+    ],
+  },
+  {
+    id: "448f69c5296fb1bbaafe5aad28c6ad52",
+    href: "/content/content-question-bank-atpl/annexes/448f69c5296fb1bbaafe5aad28c6ad52.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "Q3220RD7AF",
+        href: "/modules/[object Object]/questions/Q3220RD7AF",
+      },
+      {
+        id: "QQV4FC0AB1",
+        href: "/modules/[object Object]/questions/QQV4FC0AB1",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.04.02.01",
+        href: "/modules/[object Object]/learning-objectives/010.06.04.02.01",
+      },
+    ],
+  },
+  {
+    id: "ea013b8d7dd4c0a7cabba3b64452a666",
+    href: "/content/content-question-bank-atpl/annexes/ea013b8d7dd4c0a7cabba3b64452a666.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QN4IQGLTNJ",
+        href: "/modules/[object Object]/questions/QN4IQGLTNJ",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.04.03.02",
+        href: "/modules/[object Object]/learning-objectives/010.06.04.03.02",
+      },
+    ],
+  },
+  {
+    id: "a5d5c51419b80b4def0a4e318cda5e17",
+    href: "/content/content-question-bank-atpl/annexes/a5d5c51419b80b4def0a4e318cda5e17.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QN4IQGLTNJ",
+        href: "/modules/[object Object]/questions/QN4IQGLTNJ",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.04.03.02",
+        href: "/modules/[object Object]/learning-objectives/010.06.04.03.02",
+      },
+    ],
+  },
+  {
+    id: "e6dea6d8e7ef4336a4f309075a55b94b",
+    href: "/content/content-question-bank-atpl/annexes/e6dea6d8e7ef4336a4f309075a55b94b.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QN4IQGLTNJ",
+        href: "/modules/[object Object]/questions/QN4IQGLTNJ",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.04.03.02",
+        href: "/modules/[object Object]/learning-objectives/010.06.04.03.02",
+      },
+    ],
+  },
+  {
+    id: "d73b5e738bdb6316d05b6fd961edb6c7",
+    href: "/content/content-question-bank-atpl/annexes/d73b5e738bdb6316d05b6fd961edb6c7.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QN4IQGLTNJ",
+        href: "/modules/[object Object]/questions/QN4IQGLTNJ",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.06.04.03.02",
+        href: "/modules/[object Object]/learning-objectives/010.06.04.03.02",
+      },
+    ],
+  },
+  {
+    id: "33d52b04ff66e08e76d2ad4c1dfeb11f",
+    href: "/content/content-question-bank-atpl/annexes/33d52b04ff66e08e76d2ad4c1dfeb11f.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QBZMXBG2AT",
+        href: "/modules/[object Object]/questions/QBZMXBG2AT",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.07.01.04.06",
+        href: "/modules/[object Object]/learning-objectives/010.07.01.04.06",
+      },
+    ],
+  },
+  {
+    id: "e9026f3414f7845c5e7997322a1215fd",
+    href: "/content/content-question-bank-atpl/annexes/e9026f3414f7845c5e7997322a1215fd.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QEBWK62LU9",
+        href: "/modules/[object Object]/questions/QEBWK62LU9",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.07.01.04.06",
+        href: "/modules/[object Object]/learning-objectives/010.07.01.04.06",
+      },
+    ],
+  },
+  {
+    id: "cf69d3d9188bf359d3008dee3096b2d8",
+    href: "/content/content-question-bank-atpl/annexes/cf69d3d9188bf359d3008dee3096b2d8.jpg",
+    description: "",
+    subjects: ["010"],
+    questionBank: "atpl",
+    questions: [
+      {
+        id: "QFU5DSFC70",
+        href: "/modules/[object Object]/questions/QFU5DSFC70",
+      },
+    ],
+    learningObjectives: [
+      {
+        id: "010.07.02.15.04",
+        href: "/modules/[object Object]/learning-objectives/010.07.02.15.04",
+      },
+    ],
+  },
+];
+
+export const mockRetrieveAnnexesData: AppRouterOutput["common"]["search"]["retrieveAnnexes"] =
+  {
+    items,
+  };
+
+export const mockSearchAnnexesData: AppRouterOutput["common"]["search"]["searchAnnexes"] =
+  {
+    items,
+    totalResults: 697,
+    nextCursor: 20,
+  };
