+++ libs/next/question-bank/src/components/markdown-from-server/index.ts
@@ -1,0 +1,1 @@
+export { MarkdownFromServer } from "./markdown-from-server";
