+++ libs/next/question-bank/src/containers/question-search/question-search.mock.tsx
@@ -30,985 +30,3 @@ export const mockData: AppRouterOutput["containers"]["questions"]["getQuestionSe
       ],
     },
   };
-
-export const mockSearchData: AppRouterOutput["common"]["search"]["searchQuestions"] =
-  {
-    items: [
-      {
-        id: "K7HuwYXz",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The short period oscillation should always be heavily damped.\n- :x: When the phugoid is slightly unstable, an aeroplane becomes uncontrollable.\n- :x: The phugoid should always be heavily damped.\n- :x: A slightly unstable short period oscillation is no problem for an aeroplane.",
-        externalIds: ["BGS-810670", "BGS-812108", "AVEXAM-7288"],
-        href: "/modules/atpl/questions/K7HuwYXz",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "ezh4CuEv",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: VR must not be less than 1.05 VMCA and not less than V1.\n- :x: VR must not be less than 1.05 VMCA and not less than 1.1 V1.\n- :x: VR must not be less than 1.1 VMCA and not less than V1.\n- :x: VR must not be less than VMCA and not less than 1.05 V1.",
-        externalIds: [
-          "BGS-320653",
-          "AVEXAM-57312",
-          "ATPLGS-625808",
-          "ATPLQ-321025",
-        ],
-        href: "/modules/atpl/questions/ezh4CuEv",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "Ax7FFF3g",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Oxygen diffusion from the blood into the cells depends on the partial oxygen pressure gradient.\n- :x: The gradient of diffusion is higher at altitude than it is at sea level.\n- :x: Oxygen is mainly transported in plasma.\n- :x: Oxygen diffusion from the lungs into the blood is independent of the partial oxygen pressure.",
-        externalIds: ["ATPLGS-626755"],
-        href: "/modules/atpl/questions/Ax7FFF3g",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "SxFJ91nN",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: V~R~ must not be less than 1.05 V~MCA~ and not less than V~1~.\n- :x: V~R~ must not be less than 1.1 V~MCA ~and not less than V~1~.\n- :x: V~R~ must not be less than V~MCA~ and not less than 1.05 V~1~.\n- :x: V~R~ must not be less than 1.05 V~MCA~ and not less than 1.1 V~1~.",
-        externalIds: [
-          "ATPLQ-321025",
-          "ATPLGS-625808",
-          "AVEXAM-57312",
-          "BGS-320653",
-        ],
-        href: "/modules/atpl/questions/SxFJ91nN",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "uV7RVfAr",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Spoiler extension increases the stall speed, the minimum rate of descent and the minimum angle of descent.\n- :x: Flap extension reduces the maximum lift/drag ratio thus reducing the minimum rate of descent.\n- :x: Flap extension reduces the stall speed, which increases the maximum glide distance.\n- :x: Flap extension has no effect on the minimum rate of descent as this is only affected by TAS.",
-        externalIds: ["BGS-811633", "BGS-810466", "AVEXAM-68132"],
-        href: "/modules/atpl/questions/uV7RVfAr",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "2qQjMgG1",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: As the angle of attack increases, the stagnation point on the aerofoil's profile moves downwards on the profiles leading edge.\n- :x: The stagnation point is always situated on the chord line, the centre of pressure is not.\n- :x: The centre of pressure is the point on the aerofoil leading edge where the airflow splits up.\n- :x: The stagnation point is another name for centre of pressure.",
-        externalIds: ["AVEXAM-54692"],
-        href: "/modules/atpl/questions/2qQjMgG1",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "UrGzEwAj",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: A <strong>ridge</strong> is an elongated area of relatively high pressure, a <strong>trough</strong> is a elongated area of relatively low pressure and a <strong>col</strong> is a region between two highs and two lows.\n- :x: A <strong>ridge</strong> is an elongated area of relatively high pressure, a c<strong>ol </strong>is a elongated area of relatively low pressure and a <strong>trough</strong> is a region between two highs and two lows.\n- :x: A <strong>col </strong>is an elongated area of relatively high pressure, a <strong>trough</strong> is a elongated area of relatively low pressure and a <strong>ridge</strong> is a region between two highs and two lows.\n- :x: A <strong>trough</strong> is an elongated area of relatively high pressure, a <strong>ridge</strong> is a elongated area of relatively low pressure and a <strong>col</strong> is a region between two highs and two lows.",
-        externalIds: ["BGS-995620"],
-        href: "/modules/atpl/questions/UrGzEwAj",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "kA7q1I98",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The climb limited take-off mass depends on pressure altitude and outer air temperature.\n- :x: The performance limited take-off mass is the highest of: field length limited take-off mass climb limited take-off mass obstacle limited take-off mass.\n- :x: The climb limited take-off mass will increase if the headwind component increases.\n- :x: The climb limited take-off mass increases when a larger take-off flap setting is used.",
-        externalIds: [
-          "BGS-320656",
-          "ATPLGS-107415",
-          "ATPLGS-107575",
-          "ATPLGS-114227",
-          "ATPLQ-322985",
-          "AVEXAM-56017",
-          "AVEXAM-90038",
-          "AVEXAM-13194",
-          "AVEXAM-80302",
-          "ATPLQ-326469",
-          "ATPLGS-625115",
-          "ATPLQ-329553",
-          "ATPLQ-329630",
-          "ATPLQ-321955",
-          "ATPLGS-107438",
-          "ATPLQ-324938",
-          "AVEXAM-26125",
-        ],
-        href: "/modules/atpl/questions/kA7q1I98",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "IJKUxABS",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Fog can be supercooled and can also contain ice crystals.\n- :x: Mist and haze consist of water droplets.\n- :x: Fog and haze do not occur in the tropics.\n- :x: Mist and haze only differ by different values of visibility.",
-        externalIds: ["AVEXAM-69003", "ATPLQ-505558"],
-        href: "/modules/atpl/questions/IJKUxABS",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "fklnIxec",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The lower limit of a TMA shall be established at a height of at least 700 ft AGL.\n- :x: The upper limit of a CTR shall be established at a height of at least 3 000 ft AMSL.\n- :x: The lower limit of a CTA shall be established at a height of at least 1 500 ft AGL.\n- :x: The lower limit of an UIR may coincide with an IFR cruising level.",
-        externalIds: ["AVEXAM-21315", "ATPLGS-621989"],
-        href: "/modules/atpl/questions/fklnIxec",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "lapzBT5F",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Flap extension causes a reduction in stall speed and the maximum glide distance.\n- :x: Flap extension will increase (CL/CD)max thus causing a reduction in the minimum rate of descent.\n- :x: Flap extension has no influence on the minimum rate of descent, as only TAS has to be taken into account.\n- :x: Spoiler extension causes a reduction in stall speed and the minimum rate of descent, but increases the minimum descent angle.",
-        externalIds: ["ATPLGS-625875", "BGS-810432"],
-        href: "/modules/atpl/questions/lapzBT5F",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "BRYanzNr",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Dynamic stability is possible only when the aeroplane is statically stable about the relevant axis.\n- :x: A dynamically stable aeroplane would be almost impossible to fly manually.\n- :x: Dynamic stability about the lateral axis implies that after being displaced from its original equilibrium condition, the aeroplane will return to that condition without oscillation.\n- :x: Static stability means that the aeroplane is also dynamically stable about the relevant axis.",
-        externalIds: ["ATPLQ-814951", "BGS-810648"],
-        href: "/modules/atpl/questions/BRYanzNr",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "1sAtIFk2",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Problems in the personal relationships between crew members are likely to hamper the communication process.\n- :x: Inconsistent communication improves flight safety.\n- :x: Personal conflict that takes place prior to take-off should wait to be addressed until the end of the flight.\n- :x: There is no relation between inadequate communication and incidents or accidents.",
-        externalIds: ["ATPLQ-402286", "AVEXAM-66724", "ATPLGS-626752"],
-        href: "/modules/atpl/questions/1sAtIFk2",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "8oFqQuHL",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The Solar System consists of the Sun, planets and stars.\n- :x: The planets move around the Sun like all stars of the Solar System.\n- :x: The Earth is one of the planets which are all moving in elliptical orbit around the Sun.\n- :x: The Sun moves in an elliptical orbit around the Earth.",
-        externalIds: ["AVEXAM-10586"],
-        href: "/modules/atpl/questions/8oFqQuHL",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "BgHXCT7I",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The climb limited take-off mass depends on pressure altitude and outer air temperature.\n- :x: The climb limited take-off mass will increase if the headwind component increases.\n- :x: The performance limited take-off mass is the highest of field length limited take-off mass climb limited take-off mass obstacle limited take-off mass.\n- :x: The climb limited take-off mass increases when a larger take-off flap setting is used.",
-        externalIds: [
-          "ATPLQ-323595",
-          "ATPLGS-107415",
-          "ATPLGS-107575",
-          "ATPLGS-114227",
-          "BGS-320264",
-          "BGS-320248",
-          "BGS-320224",
-          "BGS-320656",
-          "AVEXAM-41520",
-          "AVEXAM-56017",
-          "AVEXAM-90038",
-          "AVEXAM-13194",
-          "AVEXAM-80302",
-          "ATPLGS-625115",
-          "ATPLGS-107438",
-          "AVEXAM-26125",
-        ],
-        href: "/modules/atpl/questions/BgHXCT7I",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "Y5RD4ZDw",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: During a phugoid altitude varies significantly, but during a short period oscillation it remains approximately constant.\n- :x: During both a phugoid and a short period oscillation altitude remains approximately constant.\n- :x: During both a phugoid and a short period oscillation altitude varies significantly.\n- :x: During a phugoid altitude remains approximately constant, but during a short period oscillation it varies significantly.",
-        externalIds: ["BGS-810669", "AVEXAM-17048"],
-        href: "/modules/atpl/questions/Y5RD4ZDw",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "ZeQuarz6",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: The climb limited Take-off Mass depends on pressure altitude and outside air temperature.\n- :x: The climb limited Take-off Mass increases when a larger take-off flap setting is used.\n- :x: The performance limited Take-off Mass is the highest of:\ni) field length limited Take-off Mass;\nii) climb limited Take-off Mass;\niii) obstacle limited Take-off Mass.\n- :x: The climb limited Take-off Mass will increase if the headwind component increases.",
-        externalIds: [
-          "AVEXAM-80302",
-          "ATPLGS-114227",
-          "BGS-320248",
-          "ATPLQ-322985",
-          "ATPLQ-329553",
-          "ATPLQ-321955",
-          "ATPLGS-107438",
-        ],
-        href: "/modules/atpl/questions/ZeQuarz6",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "FkDh9o1x",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Obstacle limited takeoff mass\n- :x: Climb limited takeoff mass\n- :x: Tyre speed limited takeoff mass\n- :x: Field length limited takeoff mass",
-        externalIds: [
-          "AVEXAM-13194",
-          "BGS-320215",
-          "ATPLQ-329010",
-          "ATPLGS-107455",
-        ],
-        href: "/modules/atpl/questions/FkDh9o1x",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "gid9ddHM", href: "/modules/atpl/questions/gid9ddHM" },
-        ],
-      },
-      {
-        id: "gid9ddHM",
-        questionBank: "atpl",
-        subjects: ["010", "032", "040", "050", "061", "081"],
-        text: "Which statement is correct?\n\n- :white_check_mark: Extension of flaps causes a reduction of the stall speed, the maximum glide distance also reduces.\n- :x: Extension of flaps has no influence on the minimum rate of descent, as only the TAS has to be taken into account.\n- :x: Extension of flaps will increase (CL/CD MAX), causing the minimum rate of descent to decrease.\n- :x: Spoiler extension decreases the stall speed and the minimum rate of descent, but increases the minimum descent angle.",
-        externalIds: ["AVEXAM-41717"],
-        href: "/modules/atpl/questions/gid9ddHM",
-        learningObjectives: [
-          { id: "010.07", href: "/modules/atpl/learning-objectives/010.07" },
-          { id: "032.04", href: "/modules/atpl/learning-objectives/032.04" },
-          { id: "040.02", href: "/modules/atpl/learning-objectives/040.02" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.07", href: "/modules/atpl/learning-objectives/050.07" },
-          { id: "061.01", href: "/modules/atpl/learning-objectives/061.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.04", href: "/modules/atpl/learning-objectives/081.04" },
-        ],
-        relatedQuestions: [
-          { id: "K7HuwYXz", href: "/modules/atpl/questions/K7HuwYXz" },
-          { id: "ezh4CuEv", href: "/modules/atpl/questions/ezh4CuEv" },
-          { id: "Ax7FFF3g", href: "/modules/atpl/questions/Ax7FFF3g" },
-          { id: "SxFJ91nN", href: "/modules/atpl/questions/SxFJ91nN" },
-          { id: "uV7RVfAr", href: "/modules/atpl/questions/uV7RVfAr" },
-          { id: "2qQjMgG1", href: "/modules/atpl/questions/2qQjMgG1" },
-          { id: "UrGzEwAj", href: "/modules/atpl/questions/UrGzEwAj" },
-          { id: "kA7q1I98", href: "/modules/atpl/questions/kA7q1I98" },
-          { id: "IJKUxABS", href: "/modules/atpl/questions/IJKUxABS" },
-          { id: "fklnIxec", href: "/modules/atpl/questions/fklnIxec" },
-          { id: "lapzBT5F", href: "/modules/atpl/questions/lapzBT5F" },
-          { id: "BRYanzNr", href: "/modules/atpl/questions/BRYanzNr" },
-          { id: "1sAtIFk2", href: "/modules/atpl/questions/1sAtIFk2" },
-          { id: "8oFqQuHL", href: "/modules/atpl/questions/8oFqQuHL" },
-          { id: "BgHXCT7I", href: "/modules/atpl/questions/BgHXCT7I" },
-          { id: "Y5RD4ZDw", href: "/modules/atpl/questions/Y5RD4ZDw" },
-          { id: "ZeQuarz6", href: "/modules/atpl/questions/ZeQuarz6" },
-          { id: "FkDh9o1x", href: "/modules/atpl/questions/FkDh9o1x" },
-        ],
-      },
-      {
-        id: "qAHZS9GQ",
-        questionBank: "atpl",
-        subjects: ["010", "040", "050", "070", "081"],
-        text: "Which of the following statements is true?\n\n- :white_check_mark: Mist is a reduction of visibility due to the presence of (liquid) water droplets in the air.\n- :x: Standing waves are always accompanied by severe turbulence.\n- :x: Airframe icing cannot occur in clear air.\n- :x: An isotach is a line joining all points of equal windshear.",
-        externalIds: ["BGS-501393"],
-        href: "/modules/atpl/questions/qAHZS9GQ",
-        learningObjectives: [
-          { id: "010.08", href: "/modules/atpl/learning-objectives/010.08" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.01", href: "/modules/atpl/learning-objectives/050.01" },
-          { id: "050.02", href: "/modules/atpl/learning-objectives/050.02" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.08", href: "/modules/atpl/learning-objectives/050.08" },
-          { id: "050.10", href: "/modules/atpl/learning-objectives/050.10" },
-          { id: "071.01", href: "/modules/atpl/learning-objectives/071.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.06", href: "/modules/atpl/learning-objectives/081.06" },
-        ],
-        relatedQuestions: [
-          { id: "0lrHJ0lU", href: "/modules/atpl/questions/0lrHJ0lU" },
-          { id: "L4p38s2q", href: "/modules/atpl/questions/L4p38s2q" },
-          { id: "HDxA6ffA", href: "/modules/atpl/questions/HDxA6ffA" },
-          { id: "Ks1U1MM0", href: "/modules/atpl/questions/Ks1U1MM0" },
-          { id: "KHg6zlxZ", href: "/modules/atpl/questions/KHg6zlxZ" },
-          { id: "MgOxggSM", href: "/modules/atpl/questions/MgOxggSM" },
-          { id: "fNBcOqtM", href: "/modules/atpl/questions/fNBcOqtM" },
-          { id: "rMK91GJp", href: "/modules/atpl/questions/rMK91GJp" },
-          { id: "KQPeOPbH", href: "/modules/atpl/questions/KQPeOPbH" },
-          { id: "ufkqwFph", href: "/modules/atpl/questions/ufkqwFph" },
-          { id: "PFzI3zxJ", href: "/modules/atpl/questions/PFzI3zxJ" },
-          { id: "jq6VSU7z", href: "/modules/atpl/questions/jq6VSU7z" },
-          { id: "dgWvnRpl", href: "/modules/atpl/questions/dgWvnRpl" },
-        ],
-      },
-      {
-        id: "L4p38s2q",
-        questionBank: "atpl",
-        subjects: ["010", "040", "050", "070", "081"],
-        text: "Which of the following statements is true?\n\n- :white_check_mark: Stressors accumulate thus increasing the likelihood to exhaustion.\n- :x: Stressors are independent from each other.\n- :x: Stress should always be avoided under any circumstances.\n- :x: People are capable of living without stress.",
-        externalIds: ["AVEXAM-80839"],
-        href: "/modules/atpl/questions/L4p38s2q",
-        learningObjectives: [
-          { id: "010.08", href: "/modules/atpl/learning-objectives/010.08" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.01", href: "/modules/atpl/learning-objectives/050.01" },
-          { id: "050.02", href: "/modules/atpl/learning-objectives/050.02" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.08", href: "/modules/atpl/learning-objectives/050.08" },
-          { id: "050.10", href: "/modules/atpl/learning-objectives/050.10" },
-          { id: "071.01", href: "/modules/atpl/learning-objectives/071.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.06", href: "/modules/atpl/learning-objectives/081.06" },
-        ],
-        relatedQuestions: [
-          { id: "qAHZS9GQ", href: "/modules/atpl/questions/qAHZS9GQ" },
-          { id: "0lrHJ0lU", href: "/modules/atpl/questions/0lrHJ0lU" },
-          { id: "HDxA6ffA", href: "/modules/atpl/questions/HDxA6ffA" },
-          { id: "Ks1U1MM0", href: "/modules/atpl/questions/Ks1U1MM0" },
-          { id: "KHg6zlxZ", href: "/modules/atpl/questions/KHg6zlxZ" },
-          { id: "MgOxggSM", href: "/modules/atpl/questions/MgOxggSM" },
-          { id: "fNBcOqtM", href: "/modules/atpl/questions/fNBcOqtM" },
-          { id: "rMK91GJp", href: "/modules/atpl/questions/rMK91GJp" },
-          { id: "KQPeOPbH", href: "/modules/atpl/questions/KQPeOPbH" },
-          { id: "ufkqwFph", href: "/modules/atpl/questions/ufkqwFph" },
-          { id: "PFzI3zxJ", href: "/modules/atpl/questions/PFzI3zxJ" },
-          { id: "jq6VSU7z", href: "/modules/atpl/questions/jq6VSU7z" },
-          { id: "dgWvnRpl", href: "/modules/atpl/questions/dgWvnRpl" },
-        ],
-      },
-      {
-        id: "HDxA6ffA",
-        questionBank: "atpl",
-        subjects: ["010", "040", "050", "070", "081"],
-        text: "Which of the following statements is true?\n\n- :white_check_mark: QNH can be equal to QFE.\n- :x: QNH is always equal to QFE.\n- :x: QNH is always higher than QFE.\n- :x: QNH is always lower than QFE.",
-        externalIds: ["BGS-500978", "AVEXAM-34930"],
-        href: "/modules/atpl/questions/HDxA6ffA",
-        learningObjectives: [
-          { id: "010.08", href: "/modules/atpl/learning-objectives/010.08" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.01", href: "/modules/atpl/learning-objectives/050.01" },
-          { id: "050.02", href: "/modules/atpl/learning-objectives/050.02" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.08", href: "/modules/atpl/learning-objectives/050.08" },
-          { id: "050.10", href: "/modules/atpl/learning-objectives/050.10" },
-          { id: "071.01", href: "/modules/atpl/learning-objectives/071.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.06", href: "/modules/atpl/learning-objectives/081.06" },
-        ],
-        relatedQuestions: [
-          { id: "qAHZS9GQ", href: "/modules/atpl/questions/qAHZS9GQ" },
-          { id: "0lrHJ0lU", href: "/modules/atpl/questions/0lrHJ0lU" },
-          { id: "L4p38s2q", href: "/modules/atpl/questions/L4p38s2q" },
-          { id: "Ks1U1MM0", href: "/modules/atpl/questions/Ks1U1MM0" },
-          { id: "KHg6zlxZ", href: "/modules/atpl/questions/KHg6zlxZ" },
-          { id: "MgOxggSM", href: "/modules/atpl/questions/MgOxggSM" },
-          { id: "fNBcOqtM", href: "/modules/atpl/questions/fNBcOqtM" },
-          { id: "rMK91GJp", href: "/modules/atpl/questions/rMK91GJp" },
-          { id: "KQPeOPbH", href: "/modules/atpl/questions/KQPeOPbH" },
-          { id: "ufkqwFph", href: "/modules/atpl/questions/ufkqwFph" },
-          { id: "PFzI3zxJ", href: "/modules/atpl/questions/PFzI3zxJ" },
-          { id: "jq6VSU7z", href: "/modules/atpl/questions/jq6VSU7z" },
-          { id: "dgWvnRpl", href: "/modules/atpl/questions/dgWvnRpl" },
-        ],
-      },
-      {
-        id: "Ks1U1MM0",
-        questionBank: "atpl",
-        subjects: ["010", "040", "050", "070", "081"],
-        text: "Which of the following statements is true?\n\n- :white_check_mark: Aircraft Basic Mass tends to increase slightly over\ntime\n- :x: Aircraft mass does not affect aircraft\nperformance\n- :x: Aircraft Basic Mass tends to decrease slightly over\ntime\n- :x: Aircraft Basic Mass  is always the same",
-        externalIds: ["BGS-991461"],
-        href: "/modules/atpl/questions/Ks1U1MM0",
-        learningObjectives: [
-          { id: "010.08", href: "/modules/atpl/learning-objectives/010.08" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.01", href: "/modules/atpl/learning-objectives/050.01" },
-          { id: "050.02", href: "/modules/atpl/learning-objectives/050.02" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.08", href: "/modules/atpl/learning-objectives/050.08" },
-          { id: "050.10", href: "/modules/atpl/learning-objectives/050.10" },
-          { id: "071.01", href: "/modules/atpl/learning-objectives/071.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.06", href: "/modules/atpl/learning-objectives/081.06" },
-        ],
-        relatedQuestions: [
-          { id: "qAHZS9GQ", href: "/modules/atpl/questions/qAHZS9GQ" },
-          { id: "0lrHJ0lU", href: "/modules/atpl/questions/0lrHJ0lU" },
-          { id: "L4p38s2q", href: "/modules/atpl/questions/L4p38s2q" },
-          { id: "HDxA6ffA", href: "/modules/atpl/questions/HDxA6ffA" },
-          { id: "KHg6zlxZ", href: "/modules/atpl/questions/KHg6zlxZ" },
-          { id: "MgOxggSM", href: "/modules/atpl/questions/MgOxggSM" },
-          { id: "fNBcOqtM", href: "/modules/atpl/questions/fNBcOqtM" },
-          { id: "rMK91GJp", href: "/modules/atpl/questions/rMK91GJp" },
-          { id: "KQPeOPbH", href: "/modules/atpl/questions/KQPeOPbH" },
-          { id: "ufkqwFph", href: "/modules/atpl/questions/ufkqwFph" },
-          { id: "PFzI3zxJ", href: "/modules/atpl/questions/PFzI3zxJ" },
-          { id: "jq6VSU7z", href: "/modules/atpl/questions/jq6VSU7z" },
-          { id: "dgWvnRpl", href: "/modules/atpl/questions/dgWvnRpl" },
-        ],
-      },
-      {
-        id: "KHg6zlxZ",
-        questionBank: "atpl",
-        subjects: ["010", "040", "050", "070", "081"],
-        text: "Which of the following statements is true?\n\n- :white_check_mark: VRA must be less than or equal to VB\n- :x: VRA and VB are the same thing\n- :x: VRA must be greater than VB\n- :x: VRA must be less than VB",
-        externalIds: ["BGS-991467"],
-        href: "/modules/atpl/questions/KHg6zlxZ",
-        learningObjectives: [
-          { id: "010.08", href: "/modules/atpl/learning-objectives/010.08" },
-          { id: "040.03", href: "/modules/atpl/learning-objectives/040.03" },
-          { id: "050.01", href: "/modules/atpl/learning-objectives/050.01" },
-          { id: "050.02", href: "/modules/atpl/learning-objectives/050.02" },
-          { id: "050.04", href: "/modules/atpl/learning-objectives/050.04" },
-          { id: "050.08", href: "/modules/atpl/learning-objectives/050.08" },
-          { id: "050.10", href: "/modules/atpl/learning-objectives/050.10" },
-          { id: "071.01", href: "/modules/atpl/learning-objectives/071.01" },
-          { id: "081.01", href: "/modules/atpl/learning-objectives/081.01" },
-          { id: "081.06", href: "/modules/atpl/learning-objectives/081.06" },
-        ],
-        relatedQuestions: [
-          { id: "qAHZS9GQ", href: "/modules/atpl/questions/qAHZS9GQ" },
-          { id: "0lrHJ0lU", href: "/modules/atpl/questions/0lrHJ0lU" },
-          { id: "L4p38s2q", href: "/modules/atpl/questions/L4p38s2q" },
-          { id: "HDxA6ffA", href: "/modules/atpl/questions/HDxA6ffA" },
-          { id: "Ks1U1MM0", href: "/modules/atpl/questions/Ks1U1MM0" },
-          { id: "MgOxggSM", href: "/modules/atpl/questions/MgOxggSM" },
-          { id: "fNBcOqtM", href: "/modules/atpl/questions/fNBcOqtM" },
-          { id: "rMK91GJp", href: "/modules/atpl/questions/rMK91GJp" },
-          { id: "KQPeOPbH", href: "/modules/atpl/questions/KQPeOPbH" },
-          { id: "ufkqwFph", href: "/modules/atpl/questions/ufkqwFph" },
-          { id: "PFzI3zxJ", href: "/modules/atpl/questions/PFzI3zxJ" },
-          { id: "jq6VSU7z", href: "/modules/atpl/questions/jq6VSU7z" },
-          { id: "dgWvnRpl", href: "/modules/atpl/questions/dgWvnRpl" },
-        ],
-      },
-    ],
-    totalResults: 3693,
-    nextCursor: 24,
-  };
