+++ libs/next/tests/src/components/test-question-result/test-question-result.tsx
@@ -2,7 +2,10 @@ import { forwardRef } from "react";
 import { default as CheckIcon } from "@mui/icons-material/Check";
 import { default as CrossIcon } from "@mui/icons-material/Close";
 import { Box, Sheet, Stack, Typography } from "@mui/joy";
-import { MarkdownClient, MarkdownClientCompressed } from "../markdown-client";
+import {
+  MarkdownClient,
+  MarkdownClientCompressed,
+} from "@chair-flight/react/components";
 import type { SheetProps } from "@mui/joy";
 
 export type TestQuestionResultProps = {
