+++ libs/react/components/src/index.ts
@@ -15,25 +15,22 @@ export * from "./hooks/use-media-query";
 export * from "./hooks/use-window-resize";
 export * from "./hooks/use-page-transition";
 export * from "./hooks/use-persistence";
+export * from "./hooks/use-throttle";
 export * from "./image-viewer";
 export * from "./image-with-modal";
 export * from "./input-slider";
-export * from "./learning-objectives-list";
 export * from "./loading-button";
 export * from "./markdown";
 export * from "./markdown-client";
 export * from "./module-selection-button";
 export * from "./nested-checkbox-select";
-export * from "./question-list";
 export * from "./question-multiple-choice";
 export * from "./question-navigation";
-export * from "./question-variant-preview";
 export * from "./search-filters";
 export * from "./search-header";
 export * from "./search-query";
 export * from "./search-list";
 export * from "./sidebar";
-export * from "./test-question-result";
 export * from "./toaster";
 export * from "./typical";
 export * from "./ups";
