+++ libs/react/components/src/question-multiple-choice/get-question-status-color.tsx
