--- libs/react/components/src/question-variant-preview/index.ts
@@ -1,2 +1,0 @@
-export { QuestionVariantPreview } from "./question-variant-preview";
-export type { QuestionVariantPreviewProps } from "./question-variant-preview";
