+++ libs/react/components/src/search-list/search-list.stories.tsx
@@ -1,14 +1,13 @@
-import { QuestionList } from "../question-list";
-import * as QuestionListStories from "../question-list/question-list.stories";
 import { SearchList } from "./search-list";
-import type { Meta, StoryObj } from "@storybook/react";
+import type { Meta } from "@storybook/react";
 
-export const QuestionListStory: StoryObj<typeof QuestionList> = {
-  ...QuestionListStories.Playground,
-  name: "QuestionList",
-  args: QuestionListStories.default.args,
-  render: (props) => <QuestionList {...props} />,
-};
+// TODO recover this demo....
+// export const QuestionListStory: StoryObj<typeof QuestionList> = {
+//   ...QuestionListStories.Playground,
+//   name: "QuestionList",
+//   args: QuestionListStories.default.args,
+//   render: (props) => <QuestionList {...props} />,
+// };
 
 const meta: Meta<typeof SearchList> = {
   title: "Components/SearchList",
