--- libs/react/markdown/src/common/jsx-runtime.cjs.d.ts
@@ -1,8 +1,0 @@
-declare const jsxRuntime: {
-  jsxs: unknown;
-  jsxDEV: unknown;
-  jsx: unknown;
-  jsxDev: unknown;
-};
-
-export { jsxRuntime };
