+++ libs/react/markdown/src/components/markdown/index.ts
@@ -1,0 +1,2 @@
+export { Markdown } from "./markdown";
+export type { MarkdownProps } from "./markdown";
