+++ libs/react/markdown/src/components/mdx/index.ts
@@ -1,0 +1,2 @@
+export { Mdx } from "./mdx";
+export type { MdxProps } from "./mdx";
