+++ libs/react/markdown/src/index.ts
@@ -1,1 +1,4 @@
-export { renderMarkdown } from "./markdown/markdown-compiler";
+import "@chair-flight/react/theme";
+
+export * from "./components/markdown";
+export * from "./components/mdx";
