+++ tsconfig.base.json
@@ -35,6 +35,7 @@
       "@chair-flight/core/question-bank": ["libs/core/question-bank/src/index.ts"],
       "@chair-flight/core/search": ["libs/core/search/src/index.ts"],
       "@chair-flight/core/tests": ["libs/core/tests/src/index.ts"],
+      "@chair-flight/core/markdown": ["libs/core/markdown/src/index.ts"],
       "@chair-flight/next/analytics": ["libs/next/analytics/src/index.ts"],
       "@chair-flight/next/public": ["libs/next/public/src/index.ts"],
       "@chair-flight/next/question-bank": ["libs/next/question-bank/src/index.ts"],
